requirements.txt
streamlit
langchain
faiss-cpu
openai
gemini-api
python-dotenv
PyPDF2
pandas
openpyxl
